+++ FourSquare/Controllers/DetailVenue/ViewPageImageHeader/PageImageHeaderView.xib
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="10117" systemVersion="15G31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="10085"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="PageImageHeaderView" customModule="FourSquare" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="375" height="250"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Gl7-jr-hgf">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="250"/>
+                    <subviews>
+                        <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="eHN-cd-hFI">
+                            <rect key="frame" x="0.0" y="0.0" width="375" height="250"/>
+                            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
+                        </view>
+                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Jyy-ot-ae7">
+                            <rect key="frame" x="10" y="114" width="13" height="22"/>
+                            <state key="normal" image="back_page_ic"/>
+                            <connections>
+                                <action selector="didTabBackAction:" destination="iN0-l3-epB" eventType="touchUpInside" id="ELL-A6-uDS"/>
+                            </connections>
+                        </button>
+                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="BfA-8j-Gks">
+                            <rect key="frame" x="352" y="114" width="13" height="22"/>
+                            <state key="normal" image="next_page_ic"/>
+                            <connections>
+                                <action selector="didTabNextAction:" destination="iN0-l3-epB" eventType="touchUpInside" id="Iyn-xl-xUS"/>
+                            </connections>
+                        </button>
+                        <pageControl opaque="NO" contentMode="scaleToFill" enabled="NO" contentHorizontalAlignment="center" contentVerticalAlignment="center" numberOfPages="3" translatesAutoresizingMaskIntoConstraints="NO" id="Jif-Qu-tbO">
+                            <rect key="frame" x="168" y="213" width="39" height="37"/>
+                        </pageControl>
+                    </subviews>
+                    <constraints>
+                        <constraint firstItem="eHN-cd-hFI" firstAttribute="leading" secondItem="Gl7-jr-hgf" secondAttribute="leading" id="6Sa-WY-Q6x"/>
+                        <constraint firstAttribute="bottom" secondItem="eHN-cd-hFI" secondAttribute="bottom" id="DHh-hI-OZ0"/>
+                        <constraint firstItem="eHN-cd-hFI" firstAttribute="top" secondItem="Gl7-jr-hgf" secondAttribute="top" id="IIA-1z-c3z"/>
+                        <constraint firstAttribute="trailing" secondItem="eHN-cd-hFI" secondAttribute="trailing" id="RlW-pf-579"/>
+                        <constraint firstItem="BfA-8j-Gks" firstAttribute="centerY" secondItem="eHN-cd-hFI" secondAttribute="centerY" id="Zp2-8d-JcS"/>
+                        <constraint firstAttribute="trailing" secondItem="BfA-8j-Gks" secondAttribute="trailing" constant="10" id="bvo-77-z6Y"/>
+                        <constraint firstItem="Jyy-ot-ae7" firstAttribute="leading" secondItem="Gl7-jr-hgf" secondAttribute="leading" constant="10" id="cFD-n2-tSR"/>
+                        <constraint firstAttribute="bottom" secondItem="Jif-Qu-tbO" secondAttribute="bottom" id="hmH-5u-ifz"/>
+                        <constraint firstItem="Jif-Qu-tbO" firstAttribute="centerX" secondItem="eHN-cd-hFI" secondAttribute="centerX" id="oeI-ss-paq"/>
+                        <constraint firstItem="Jyy-ot-ae7" firstAttribute="centerY" secondItem="eHN-cd-hFI" secondAttribute="centerY" id="sbq-oz-cqD"/>
+                    </constraints>
+                </view>
+            </subviews>
+            <constraints>
+                <constraint firstAttribute="bottom" secondItem="Gl7-jr-hgf" secondAttribute="bottom" id="Lnk-yM-S2L"/>
+                <constraint firstAttribute="trailing" secondItem="Gl7-jr-hgf" secondAttribute="trailing" id="ZPi-en-rLI"/>
+                <constraint firstItem="Gl7-jr-hgf" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="dLj-N5-wRp"/>
+                <constraint firstItem="Gl7-jr-hgf" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="hSZ-A7-0al"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <connections>
+                <outlet property="afterPageButton" destination="BfA-8j-Gks" id="8QK-qt-zbS"/>
+                <outlet property="beforePageButton" destination="Jyy-ot-ae7" id="2Yx-Cw-DQZ"/>
+                <outlet property="imagesPageControl" destination="Jif-Qu-tbO" id="BXm-4x-gtT"/>
+                <outlet property="imagesPageView" destination="eHN-cd-hFI" id="jNn-Sy-Q3L"/>
+                <outlet property="view" destination="Gl7-jr-hgf" id="HhU-2w-16H"/>
+            </connections>
+        </view>
+    </objects>
+    <resources>
+        <image name="back_page_ic" width="13" height="22"/>
+        <image name="next_page_ic" width="13" height="22"/>
+    </resources>
+</document>
